Move taker's `ActorSystem` to `lib` as `Taker`
